Add initial Image Manifest specification
This specifies the base format of image manifests, and a file system
layout for image stores (and local caches, which expose their store
contents via a read-only shared filesystem).

We change the image identifier from a UUID to the image manifest's
SHA256-digest. For now, we simply hard-code one arbitrary digest in
the CLI connector.

This image manifest format is currently being used for the WIP QEMU
supervisor component.