Remove `switchboard-{traits,macros}` helper crates
Having the `treadmill-rs` base crate depend on these crates means
that, if we want to publish the `treadmill-rs` crate or anything that
depends on it on `crates.io`, then we also have to publish these
helpers. For the value that these bring us, that does not seem worth
the tradeoff. We can move the single `JsonProxiedStatus` trait into
`treadmill-rs` itself and manually implement it on the required
types. If that becomes too burdensome, we can use a `macro_rules!`
expression to ease this in the future without introducing a procedural
macro.