Unify crate name prefix to "treadmill-"
We are about to publish an initial version of the Treadmill CLI client
to crates.io. While the CLI client crate is named "tml-cli", the
base-crate containing API and type definitions, which we have to
publish alongside it, is called "treadmill-rs". The "tml-" prefix,
while nice and short, is rather ambiguous. We should publish the
crates prefixed with "treadmill-", and then produce binaries that are
available under the shortened "tml" prefix (such as the CLI crate,
which produces a binary named `tml`).